changed names for decrypt and encrypt class vs runner
