+++ decrypted.txt
@@ -1,1 +1,1 @@
-{:decryption=>"hello world", :key=>"87305", :date=>"070822"}
+{:decryption=>"hello world", :key=>"47028", :date=>"070822"}
