+++ encrypted.txt
@@ -1,1 +1,1 @@
-scwuzygxbjo
+f vqmvftpgn
