+++ lib/decrypt_runner.rb
@@ -1,27 +1,6 @@
-class DecryptRunner
-  attr_reader :read_file_path,
-              :write_file_path,
-              :key,
-              :date
+require_relative './enigma'
+require_relative './decrypt'
 
-  def initialize(read_file_path, write_file_path, key, date)
-    @read_file_path = read_file_path
-    @write_file_path = write_file_path
-    @key = key
-    @date = date
-    @enigma = Enigma.new
-  end
+runner = Decrypt.new(ARGV[0], ARGV[1], ARGV[2], ARGV[3])
 
-  def run
-    encrypted = File.open(@read_file_path, "r")
-    message = encrypted.read
-    encrypted.close
-    
-    decrypted_text = @enigma.decrypt(message, key = (@key), date = (@date))
-    decrypt = File.open(@write_file_path, "w")
-    decrypt.write(decrypted_text)
-    puts "created 'decrypted.txt' with the key #{key} and date #{date}"
-    
-    decrypt.close
-  end
-end
+runner.run
