+++ lib/encrypt.rb
@@ -1,24 +1,25 @@
-require_relative './enigma'
-require_relative './encrypt_runner'
+class Encrypt
+  attr_reader :read_file_path,
+              :write_file_path
 
-runner = EncryptRunner.new(ARGV[0], ARGV[1])
+  def initialize(read_file_path, write_file_path)
+    @read_file_path = read_file_path
+    @write_file_path = write_file_path
+    @enigma = Enigma.new
+  end
 
-runner.run
+  def run
+    message_file = File.open(read_file_path, "r")
+    message = message_file.read
+    message_file.close
 
+    encrypted_text = @enigma.encrypt(message, key = @enigma.key_generator, date = @enigma.date_generator)
 
+    encrypt = File.open(write_file_path, "w")
 
+    encrypt.write(encrypted_text[:encryption])
+    puts "created 'encrypted.txt' with the key #{key} and date #{date}"
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+    encrypt.close
+  end
+end
